Extracting Submission Export to a separate feature
