+++ app/(main)/forms/[formId]/submissions/ui/submissions-table.tsx
@@ -1,11 +1,8 @@
 "use client";
 
-import { Spinner } from "@/components/loaders/spinner";
-import { Button } from "@/components/ui/button";
-import { toast } from "@/components/ui/toast";
+import { ExportSubmissionsButton } from "@/features/submissions/ui/export";
 import { COLUMNS_DEFINITION, DataTable } from "@/features/submissions/ui/table";
 import { Submission } from "@/types";
-import { Download } from "lucide-react";
 import { useEffect, useState } from "react";
 
 type SubmissionsTableProps = {
@@ -14,71 +11,9 @@ type SubmissionsTableProps = {
 };
 
 const SubmissionsTable = ({ data, formId }: SubmissionsTableProps) => {
-  const [isExporting, setIsExporting] = useState(false);
   const [selectedSubmissionId, setSelectedSubmissionId] = useState<
     string | null
   >(null);
-
-  const handleExport = async () => {
-    try {
-      setIsExporting(true);
-      
-      // Notify user immediately that export is starting
-      toast.info({
-        title: "Starting export",
-        description: "Preparing your file for download...",
-      });
-
-      // Use fetch to track the request status
-      const exportUrl = `/api/forms/${formId}/export?format=csv`;
-      
-      const response = await fetch(exportUrl);
-      
-      if (!response.ok) {
-        throw new Error(`Export failed with status: ${response.status}`);
-      }
-      
-      // Get content info from response headers
-      const contentDisposition = response.headers.get("Content-Disposition") || "";
-      let filename = `form-${formId}-submissions.csv`;
-      
-      // Try to extract filename from Content-Disposition header
-      const filenameMatch = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/.exec(contentDisposition);
-      if (filenameMatch && filenameMatch[1]) {
-        filename = filenameMatch[1].replace(/['"]/g, "");
-      }
-      
-      // Create blob from response and trigger download
-      const blob = await response.blob();
-      const url = window.URL.createObjectURL(blob);
-      const link = document.createElement("a");
-      link.href = url;
-      link.setAttribute("download", filename);
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      
-      // Clean up object URL
-      window.URL.revokeObjectURL(url);
-      
-      // Now we can safely show success message
-      toast.success({
-        title: "Export successful",
-        description: "Your file has been downloaded successfully.",
-      });
-    } catch (error) {
-      console.error("Export failed:", error);
-      toast.error({
-        title: "Export failed",
-        description:
-          typeof error === "object" && error !== null && "message" in error
-            ? String(error.message)
-            : "There was a problem exporting the submissions.",
-      });
-    } finally {
-      setIsExporting(false);
-    }
-  };
 
   useEffect(() => {
     const handleKeyDown = (e: KeyboardEvent) => {
@@ -110,14 +45,7 @@ const SubmissionsTable = ({ data, formId }: SubmissionsTableProps) => {
   return (
     <>
       <div className="flex justify-end mb-4">
-        <Button variant="outline" onClick={handleExport} disabled={isExporting}>
-          {isExporting ? (
-            <Spinner className="h-4 w-4 mr-2" />
-          ) : (
-            <Download className="h-4 w-4 mr-2" />
-          )}
-          {isExporting ? "Exporting..." : "Export Submissions"}
-        </Button>
+        <ExportSubmissionsButton formId={formId} />
       </div>
       <DataTable data={data} columns={COLUMNS_DEFINITION} />
     </>
