+++ features/submissions/ui/export/index.ts
@@ -1,0 +1,1 @@
+export { ExportSubmissionsButton } from "./export-submissions-button";
